requirements.txt
ffmpeg-python
yt-dlp
pytesseract
pillow
